Moves workflow-ui/backstack-* into workflow-ui/container-*
Because the relationship between modals and backstack is becoming circular.
