+++ samples/tictactoe/common/build.gradle.kts
@@ -5,7 +5,6 @@ plugins {
 
 dependencies {
   implementation(project(":samples:containers:common"))
-  implementation(project(":workflow-ui:backstack-common"))
   implementation(project(":workflow-ui:container-common"))
   implementation(project(":workflow-core"))
   implementation(project(":workflow-rx2"))
