+++ samples/todo-android/common/build.gradle.kts
@@ -5,7 +5,7 @@ plugins {
 
 dependencies {
   implementation(project(":samples:containers:common"))
-  implementation(project(":workflow-ui:backstack-common"))
+  implementation(project(":workflow-ui:container-common"))
   implementation(project(":workflow-core"))
   implementation(project(":workflow-rx2"))
   implementation(Dependencies.Kotlin.Stdlib.jdk6)
