--- workflow-ui/backstack-android/gradle.properties
@@ -1,3 +1,0 @@
-POM_ARTIFACT_ID=workflow-ui-backstack-android
-POM_NAME=Workflow UI Back Stack Android
-POM_PACKAGING=aar
