--- workflow-ui/backstack-android/src/androidTest/AndroidManifest.xml
@@ -1,9 +1,0 @@
-<manifest package="com.squareup.workflow1.ui.backstack.test"
-    xmlns:android="http://schemas.android.com/apk/res/android">
-
-  <application>
-    <activity
-        android:name=".fixtures.BackStackContainerLifecycleActivity"
-        android:theme="@style/Theme.AppCompat.NoActionBar"/>
-  </application>
-</manifest>
