--- workflow-ui/backstack-android/src/main/AndroidManifest.xml
@@ -1,1 +1,0 @@
-<manifest package="com.squareup.workflow1.ui.backstack"/>
