--- workflow-ui/backstack-android/src/main/res/values/ids.xml
@@ -1,12 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<resources>
-  <!--
-  BackStackContainer restricts animation to child with this id, if there is one.
-  Otherwise animates its entire body.
-   -->
-  <item type="id" name="back_stack_body"/>
-  <!-- View Tag holding the BackStackScreen.Key associated with a view. -->
-  <item type="id" name="workflow_back_stack_key"/>
-  <!-- ID used by BackStackContainer to opt in to view persistence. -->
-  <item type="id" name="workflow_back_stack_container"/>
-</resources>
