--- workflow-ui/backstack-common/gradle.properties
@@ -1,3 +1,0 @@
-POM_ARTIFACT_ID=workflow-ui-backstack-common-jvm
-POM_NAME=Workflow UI Back Stack
-POM_PACKAGING=jar
