+++ workflow-ui/container-android/src/androidTest/java/com/squareup/workflow1/ui/backstack/test/fixtures/NoTransitionBackStackContainer.kt
@@ -8,8 +8,8 @@ import com.squareup.workflow1.ui.ViewFactory
 import com.squareup.workflow1.ui.WorkflowUiExperimentalApi
 import com.squareup.workflow1.ui.backstack.BackStackContainer
 import com.squareup.workflow1.ui.backstack.BackStackScreen
-import com.squareup.workflow1.ui.backstack.R
 import com.squareup.workflow1.ui.bindShowRendering
+import com.squareup.workflow1.ui.container.R
 
 /**
  * A subclass of [BackStackContainer] that disables transitions to make it simpler to test the
