+++ workflow-ui/container-android/src/main/AndroidManifest.xml
@@ -1,1 +1,1 @@
-<manifest package="com.squareup.workflow1.ui.modal"/>
+<manifest package="com.squareup.workflow1.ui.container"/>
