+++ workflow-ui/container-android/src/main/res/layout/view_stack_layout.xml
