+++ workflow-ui/container-common/api/container-common.api
@@ -1,3 +1,23 @@
+public final class com/squareup/workflow1/ui/backstack/BackStackScreen {
+	public fun <init> (Ljava/lang/Object;Ljava/util/List;)V
+	public fun <init> (Ljava/lang/Object;[Ljava/lang/Object;)V
+	public fun equals (Ljava/lang/Object;)Z
+	public final fun get (I)Ljava/lang/Object;
+	public final fun getBackStack ()Ljava/util/List;
+	public final fun getFrames ()Ljava/util/List;
+	public final fun getTop ()Ljava/lang/Object;
+	public fun hashCode ()I
+	public final fun map (Lkotlin/jvm/functions/Function1;)Lcom/squareup/workflow1/ui/backstack/BackStackScreen;
+	public final fun mapIndexed (Lkotlin/jvm/functions/Function2;)Lcom/squareup/workflow1/ui/backstack/BackStackScreen;
+	public final fun plus (Lcom/squareup/workflow1/ui/backstack/BackStackScreen;)Lcom/squareup/workflow1/ui/backstack/BackStackScreen;
+	public fun toString ()Ljava/lang/String;
+}
+
+public final class com/squareup/workflow1/ui/backstack/BackStackScreenKt {
+	public static final fun toBackStackScreen (Ljava/util/List;)Lcom/squareup/workflow1/ui/backstack/BackStackScreen;
+	public static final fun toBackStackScreenOrNull (Ljava/util/List;)Lcom/squareup/workflow1/ui/backstack/BackStackScreen;
+}
+
 public final class com/squareup/workflow1/ui/modal/AlertContainerScreen : com/squareup/workflow1/ui/modal/HasModals {
 	public fun <init> (Ljava/lang/Object;Lcom/squareup/workflow1/ui/modal/AlertScreen;)V
 	public fun <init> (Ljava/lang/Object;Ljava/util/List;)V
